moved send.go to loaders package
