+++ internal/controller/discovery/loaders/send.go
@@ -1,12 +1,14 @@
-package core
+package loaders
 
 import (
 	"context"
 	"fmt"
+
+	"github.com/gnmic/operator/internal/controller/discovery/core"
 )
 
 // sendMessages sends discovery messages over a channel in a context-aware manner
-func sendMessages(ctx context.Context, out chan<- []DiscoveryMessage, messages []DiscoveryMessage) error {
+func sendMessages(ctx context.Context, out chan<- []core.DiscoveryMessage, messages []core.DiscoveryMessage) error {
 	select {
 	case <-ctx.Done():
 		return ctx.Err()
@@ -30,14 +32,14 @@ func forEachChunk(total, chunkSize int, fn func(start, end int) error) error {
 }
 
 // createDiscoverySnapshots takes a list of discovered targets and returns chunked DiscoverySnapshots
-func createDiscoverySnapshots(targets []DiscoveredTarget, snapshotID string, chunkSize int) []DiscoverySnapshot {
-	var snapshots []DiscoverySnapshot
+func createDiscoverySnapshots(targets []core.DiscoveredTarget, snapshotID string, chunkSize int) []core.DiscoverySnapshot {
+	var snapshots []core.DiscoverySnapshot
 	totalTargets := len(targets)
 	totalChunks := (totalTargets + chunkSize - 1) / chunkSize
 
 	_ = forEachChunk(totalTargets, chunkSize, func(i, end int) error {
 		chunk := targets[i:end]
-		snapshots = append(snapshots, DiscoverySnapshot{
+		snapshots = append(snapshots, core.DiscoverySnapshot{
 			Targets:     chunk,
 			SnapshotID:  snapshotID,
 			ChunkIndex:  i / chunkSize,
@@ -50,7 +52,7 @@ func createDiscoverySnapshots(targets []DiscoveredTarget, snapshotID string, chu
 }
 
 // SendSnapshot sends discovered targets as a snapshot over a channel in chunks
-func SendSnapshot(ctx context.Context, out chan<- []DiscoveryMessage, targets []DiscoveredTarget, snapshotID string, chunkSize int) error {
+func SendSnapshot(ctx context.Context, out chan<- []core.DiscoveryMessage, targets []core.DiscoveredTarget, snapshotID string, chunkSize int) error {
 	if len(targets) == 0 {
 		return fmt.Errorf("no targets in Snapshot")
 	}
@@ -58,7 +60,7 @@ func SendSnapshot(ctx context.Context, out chan<- []DiscoveryMessage, targets []
 	snapshots := createDiscoverySnapshots(targets, snapshotID, chunkSize)
 	for _, snapshot := range snapshots {
 		// Convert DiscoverySnapshot to DiscoveryMessage
-		messages := make([]DiscoveryMessage, 1)
+		messages := make([]core.DiscoveryMessage, 1)
 		messages[0] = snapshot
 
 		if err := sendMessages(ctx, out, messages); err != nil {
@@ -69,8 +71,8 @@ func SendSnapshot(ctx context.Context, out chan<- []DiscoveryMessage, targets []
 	return nil
 }
 
-func eventsToMessages(events []DiscoveryEvent) []DiscoveryMessage {
-	message := make([]DiscoveryMessage, len(events))
+func eventsToMessages(events []core.DiscoveryEvent) []core.DiscoveryMessage {
+	message := make([]core.DiscoveryMessage, len(events))
 	for i, event := range events {
 		message[i] = event
 	}
@@ -78,7 +80,7 @@ func eventsToMessages(events []DiscoveryEvent) []DiscoveryMessage {
 }
 
 // SendEvents sends discovery messages over channel in a context-aware manner
-func SendEvents(ctx context.Context, out chan<- []DiscoveryMessage, events []DiscoveryEvent, chunkSize int) error {
+func SendEvents(ctx context.Context, out chan<- []core.DiscoveryMessage, events []core.DiscoveryEvent, chunkSize int) error {
 	if len(events) == 0 {
 		return fmt.Errorf("no events to process")
 	}
